+++ src/test/java/org/cicirello/search/SolutionCostPairTests.java
@@ -1,6 +1,6 @@
 /*
  * Chips-n-Salsa: A library of parallel self-adaptive local search algorithms.
- * Copyright (C) 2002-2022 Vincent A. Cicirello
+ * Copyright (C) 2002-2023 Vincent A. Cicirello
  *
  * This file is part of Chips-n-Salsa (https://chips-n-salsa.cicirello.org/).
  *
@@ -67,6 +67,8 @@ public class SolutionCostPairTests {
         new SolutionCostPair<TestCopyable>(new TestCopyable(5), 3, false);
     SolutionCostPair<TestCopyable> p6 =
         new SolutionCostPair<TestCopyable>(new TestCopyable(6), 7, false);
+    SolutionCostPair<TestCopyable> pEqual =
+        new SolutionCostPair<TestCopyable>(new TestCopyable(1), 5, false);
     assertEquals(0, p1.compareTo(p2));
     assertEquals(0, p2.compareTo(p1));
     assertEquals(1, p1.compareTo(p3));
@@ -77,12 +79,51 @@ public class SolutionCostPairTests {
     assertEquals(-2, p5.compareTo(p1));
     assertEquals(-2, p1.compareTo(p6));
     assertEquals(2, p6.compareTo(p1));
+    assertEquals(0, p1.compareTo(pEqual));
+    assertEquals(0, pEqual.compareTo(p1));
+    assertEquals(p1, pEqual);
+    assertEquals(pEqual, p1);
+    assertEquals(p1.hashCode(), pEqual.hashCode());
+    assertNotEquals(p1, p2);
+    assertNotEquals(p1, p3);
+    assertNotEquals(p1, p4);
+    assertNotEquals(p1, p5);
+    assertNotEquals(p1, p6);
+    assertNotEquals(p2, p1);
+    assertNotEquals(p2, p3);
+    assertNotEquals(p2, p4);
+    assertNotEquals(p2, p5);
+    assertNotEquals(p2, p6);
+    assertNotEquals(p3, p1);
+    assertNotEquals(p3, p2);
+    assertNotEquals(p3, p4);
+    assertNotEquals(p3, p5);
+    assertNotEquals(p3, p6);
+    assertNotEquals(p4, p1);
+    assertNotEquals(p4, p2);
+    assertNotEquals(p4, p3);
+    assertNotEquals(p4, p5);
+    assertNotEquals(p4, p6);
+    assertNotEquals(p5, p1);
+    assertNotEquals(p5, p2);
+    assertNotEquals(p5, p3);
+    assertNotEquals(p5, p4);
+    assertNotEquals(p5, p6);
+    assertNotEquals(p6, p1);
+    assertNotEquals(p6, p2);
+    assertNotEquals(p6, p3);
+    assertNotEquals(p6, p4);
+    assertNotEquals(p6, p5);
+    assertNotEquals(p1, "Hello");
+    assertNotEquals(p1, new SolutionCostPair<TestCopyable>(new TestCopyable(1), 5.0, false));
+    assertNotEquals(p1, new SolutionCostPair<TestCopyable>(new TestCopyable(1), 5, true));
     p1 = new SolutionCostPair<TestCopyable>(new TestCopyable(1), 5.0, false);
     p2 = new SolutionCostPair<TestCopyable>(new TestCopyable(2), 5.0, false);
     p3 = new SolutionCostPair<TestCopyable>(new TestCopyable(3), 4.0, false);
     p4 = new SolutionCostPair<TestCopyable>(new TestCopyable(4), 6.0, false);
     p5 = new SolutionCostPair<TestCopyable>(new TestCopyable(5), 3.0, false);
     p6 = new SolutionCostPair<TestCopyable>(new TestCopyable(6), 7.0, false);
+    pEqual = new SolutionCostPair<TestCopyable>(new TestCopyable(1), 5.0, false);
     assertEquals(0, p1.compareTo(p2));
     assertEquals(0, p2.compareTo(p1));
     assertEquals(1, p1.compareTo(p3));
@@ -93,6 +134,43 @@ public class SolutionCostPairTests {
     assertEquals(-1, p5.compareTo(p1));
     assertEquals(-1, p1.compareTo(p6));
     assertEquals(1, p6.compareTo(p1));
+    assertEquals(0, p1.compareTo(pEqual));
+    assertEquals(0, pEqual.compareTo(p1));
+    assertEquals(p1, pEqual);
+    assertEquals(pEqual, p1);
+    assertEquals(p1.hashCode(), pEqual.hashCode());
+    assertNotEquals(p1, p2);
+    assertNotEquals(p1, p3);
+    assertNotEquals(p1, p4);
+    assertNotEquals(p1, p5);
+    assertNotEquals(p1, p6);
+    assertNotEquals(p2, p1);
+    assertNotEquals(p2, p3);
+    assertNotEquals(p2, p4);
+    assertNotEquals(p2, p5);
+    assertNotEquals(p2, p6);
+    assertNotEquals(p3, p1);
+    assertNotEquals(p3, p2);
+    assertNotEquals(p3, p4);
+    assertNotEquals(p3, p5);
+    assertNotEquals(p3, p6);
+    assertNotEquals(p4, p1);
+    assertNotEquals(p4, p2);
+    assertNotEquals(p4, p3);
+    assertNotEquals(p4, p5);
+    assertNotEquals(p4, p6);
+    assertNotEquals(p5, p1);
+    assertNotEquals(p5, p2);
+    assertNotEquals(p5, p3);
+    assertNotEquals(p5, p4);
+    assertNotEquals(p5, p6);
+    assertNotEquals(p6, p1);
+    assertNotEquals(p6, p2);
+    assertNotEquals(p6, p3);
+    assertNotEquals(p6, p4);
+    assertNotEquals(p6, p5);
+    assertNotEquals(p1, new SolutionCostPair<TestCopyable>(new TestCopyable(1), 5, false));
+    assertNotEquals(p1, new SolutionCostPair<TestCopyable>(new TestCopyable(1), 5.0, true));
   }
 
   private static class TestCopyable implements Copyable<TestCopyable> {
@@ -111,6 +189,11 @@ public class SolutionCostPairTests {
     @Override
     public boolean equals(Object other) {
       return other != null && ((TestCopyable) other).a == a;
+    }
+
+    @Override
+    public int hashCode() {
+      return a;
     }
   }
 }